move types into own file
