+++ src/main.ts
@@ -1,6 +1,6 @@
 import * as core from '@actions/core'
-//import { wait } from './wait.js'
-import { collectData, DataCollectOptions } from './data.js'
+import { collectData } from './data.js'
+import { AuthType, DataCollectOptions } from './types.js'
 
 /**
  * The main function for the action.
@@ -13,13 +13,17 @@ export async function run(): Promise<void> {
       org: core.getInput('org'),
       api_url: core.getInput('api_url'),
       token: core.getInput('token'),
-      auth_type: core.getInput('auth_type'),
+      auth_type: core.getInput('auth_type', {
+        required: true
+      }) as AuthType,
       is_debug: core.getInput('is_debug') === 'true' ? true : false,
       client_id: core.getInput('client_id'),
       client_secret: core.getInput('client_secret'),
       app_id: core.getInput('app_id'),
       app_private_key: core.getInput('app_private_key'),
-      app_installation_id: core.getInput('app_installation_id')
+      app_installation_id: core.getInput('app_installation_id'),
+      include_hooks: core.getInput('include_hooks') === 'true' ? true : false,
+      output_csv: core.getInput('output_csv') === 'true' ? true : false
     }
 
     // Debug logs are only output if the `ACTIONS_STEP_DEBUG` secret is true
